+++ leat/store/core/doc_store.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import Collection, Optional, Union
 
 from . import Document, DocFile
-from ..reader import TextReader, PDFReader
+from ..reader import TextReader, PDFReader, DOCXReader, PPTXReader
 
 
 class DocStore:
@@ -18,7 +18,7 @@ class DocStore:
       `ds = DocStore(LocalFileSys(filepath))`
     """
 
-    supported_file_types = {"text", "pdf"}
+    supported_file_types = {"text", "pdf", "docx", "pptx"}
     """File types supported by the document store"""
 
     def __init__(
@@ -64,6 +64,12 @@ class DocStore:
                 elif filetype == "pdf":
                     with DocFile(file, filetype=filetype).open_file(mode="rb") as ifp:
                         text = PDFReader().read(ifp)
+                elif filetype == "docx":
+                    filepath = DocFile(file, filetype=filetype).get_file()
+                    text = DOCXReader().read_file(filepath)
+                elif filetype == "pptx":
+                    filepath = DocFile(file, filetype=filetype).get_file()
+                    text = PPTXReader().read_file(filepath)
                 else:
                     print("DEBUG:", "Unknown filetype", filetype, "for", file)
             except OSError: